requirements.txt
deepgram-sdk==2.12.0
deepgram-captions